cogs/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.4.0
ffmpeg==1.4
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.11.1
yt-dlp==2025.2.19